pulled in modified bozzo's patch for specifying parameters for base types; fixed type propogation to deal with same circumstance; made number into numeric<T> and updated signatures so that types propogate properly

+++ quark/test/emit/booleans/java/Functions.java
@@ -16,7 +16,7 @@ public class Functions {
             if ((count) > (3)) {
                 return;
             }
-            count = (Integer) ((count) + (1));
+            count = (count) + (1);
         }
     }
     public static void main(String[] args) {
