+++ quark/test/emit/int-methods/java/Test.java
@@ -2,9 +2,9 @@ public class Test {
     public void test() {
         Integer x = 1;
         Integer y = 2;
-        Integer z = (Integer) ((((x) + (y)) - (3)) * (4));
+        Integer z = (((x) + (y)) - (3)) * (4);
         System.out.println(z);
-        Integer four = (Integer) ((2) + (2));
+        Integer four = (2) + (2);
         System.out.println(four);
     }
 }
